requirements.txt
pytesseract
#scikit-image
python-dateutil
requests
opencv-python
numpy
matplotlib